requirements.txt
SpeechRecognition
pyAudio
gTTS
pyttsx3
webbrowser
python-dotenv
openai